refactor: use UniFFI remote types for EthereumDecoder
- Convert EthereumDecoder to use #[uniffi::remote(Record)] pattern
- Reference gem_evm::call_decoder types directly instead of duplicating
- Simplify decode_call_internal to directly return gem_evm types
- Improve type safety and maintainability at FFI boundary
