+++ gemstone/Cargo.toml
@@ -49,8 +49,6 @@ serde_urlencoded.workspace = true
 async-trait.workspace = true
 alloy-primitives.workspace = true
 alloy-sol-types.workspace = true
-alloy-dyn-abi.workspace = true
-alloy-json-abi.workspace = true
 hex.workspace = true
 url.workspace = true
 num-bigint.workspace = true
