+++ gemstone/src/ethereum/decoder.rs
@@ -1,81 +1,28 @@
 use crate::GemstoneError;
-use alloy_dyn_abi::{DynSolValue, JsonAbiExt};
-use alloy_json_abi::JsonAbi;
-use alloy_primitives::hex;
-use alloy_sol_types::SolInterface;
-use anyhow::{anyhow, Result};
-use gem_evm::erc20::IERC20::IERC20Calls;
+use gem_evm::call_decoder;
+
+pub type GemDecodedCall = call_decoder::DecodedCall;
+pub type GemDecodedCallParam = call_decoder::DecodedCallParam;
+
+#[uniffi::remote(Record)]
+pub struct GemDecodedCall {
+    pub function: String,
+    pub params: Vec<GemDecodedCallParam>,
+}
+
+#[uniffi::remote(Record)]
+pub struct GemDecodedCallParam {
+    pub name: String,
+    pub r#type: String,
+    pub value: String,
+}
 
 #[derive(Debug, Default, uniffi::Object)]
 pub struct EthereumDecoder;
 
 impl EthereumDecoder {
-    pub fn decode_call_internal(calldata: &str, abi: Option<&str>) -> Result<DecodedCall> {
-        let calldata = hex::decode(calldata)?;
-
-        // Check minimum calldata length early
-        if calldata.len() < 4 {
-            return Err(anyhow!("Calldata too short"));
-        }
-
-        // Try ERC20 interface first if no ABI provided
-        if abi.is_none() {
-            if let Ok(call) = IERC20Calls::abi_decode(&calldata) {
-                return Ok(call.into());
-            }
-        }
-
-        if let Some(abi_str) = abi {
-            let abi = serde_json::from_str::<JsonAbi>(abi_str)?;
-            let selector = &calldata[..4];
-
-            for function in abi.functions() {
-                if function.selector() == selector {
-                    if let Ok(params) = function.abi_decode_input(&calldata[4..]) {
-                        return Ok(DecodedCall {
-                            function: function.name.clone(),
-                            params: function
-                                .inputs
-                                .iter()
-                                .zip(params.iter())
-                                .map(|(input, output)| DecodedCallParam {
-                                    name: input.name.clone(),
-                                    r#type: input.ty.to_string(),
-                                    value: Self::format_param_value(output),
-                                })
-                                .collect(),
-                        });
-                    } else {
-                        return Err(anyhow!("Failed to decode function parameters for {}", function.name));
-                    }
-                }
-            }
-            return Err(anyhow!("No matching function found for selector {:02x?}", selector));
-        }
-
-        Err(anyhow!("Failed to decode calldata"))
-    }
-
-    fn format_param_value(value: &DynSolValue) -> String {
-        use alloy_dyn_abi::DynSolValue;
-        match value {
-            DynSolValue::Address(addr) => addr.to_string(),
-            DynSolValue::Uint(val, _) => val.to_string(),
-            DynSolValue::Int(val, _) => val.to_string(),
-            DynSolValue::Bool(val) => val.to_string(),
-            DynSolValue::Bytes(val) => format!("0x{}", hex::encode(val)),
-            DynSolValue::FixedBytes(val, _) => format!("0x{}", hex::encode(val)),
-            DynSolValue::String(val) => val.clone(),
-            DynSolValue::Array(vals) | DynSolValue::FixedArray(vals) => {
-                let formatted: Vec<String> = vals.iter().map(Self::format_param_value).collect();
-                format!("[{}]", formatted.join(", "))
-            }
-            DynSolValue::Tuple(vals) => {
-                let formatted: Vec<String> = vals.iter().map(Self::format_param_value).collect();
-                format!("({})", formatted.join(", "))
-            }
-            _ => format!("{value:?}"),
-        }
+    pub fn decode_call_internal(calldata: &str, abi: Option<&str>) -> anyhow::Result<GemDecodedCall> {
+        call_decoder::decode_call(calldata, abi)
     }
 }
 
@@ -86,64 +33,11 @@ impl EthereumDecoder {
         Self {}
     }
 
-    pub fn decode_call(&self, calldata: String, abi: Option<String>) -> Result<DecodedCall, GemstoneError> {
+    pub fn decode_call(&self, calldata: String, abi: Option<String>) -> Result<GemDecodedCall, GemstoneError> {
         Self::decode_call_internal(&calldata, abi.as_deref()).map_err(GemstoneError::from)
     }
 }
 
-#[derive(Debug, PartialEq, uniffi::Record)]
-pub struct DecodedCallParam {
-    pub name: String,
-    pub r#type: String,
-    pub value: String,
-}
-
-#[derive(Debug, PartialEq, uniffi::Record)]
-pub struct DecodedCall {
-    pub function: String,
-    pub params: Vec<DecodedCallParam>,
-}
-
-impl From<IERC20Calls> for DecodedCall {
-    fn from(call: IERC20Calls) -> Self {
-        let (function, params) = match call {
-            IERC20Calls::transfer(transfer) => (
-                "transfer",
-                vec![
-                    ("to", "address", transfer.to.to_string()),
-                    ("value", "uint256", transfer.value.to_string()),
-                ]
-            ),
-            IERC20Calls::transferFrom(transfer_from) => (
-                "transferFrom", 
-                vec![
-                    ("from", "address", transfer_from.from.to_string()),
-                    ("to", "address", transfer_from.to.to_string()),
-                    ("value", "uint256", transfer_from.value.to_string()),
-                ]
-            ),
-            IERC20Calls::approve(approve) => (
-                "approve",
-                vec![
-                    ("spender", "address", approve.spender.to_string()),
-                    ("value", "uint256", approve.value.to_string()),
-                ]
-            ),
-            _ => todo!(),
-        };
-
-        DecodedCall {
-            function: function.to_string(),
-            params: params.into_iter()
-                .map(|(name, r#type, value)| DecodedCallParam {
-                    name: name.to_string(),
-                    r#type: r#type.to_string(),
-                    value,
-                })
-                .collect(),
-        }
-    }
-}
 
 #[cfg(test)]
 mod tests {
@@ -209,7 +103,7 @@ mod tests {
     #[test]
     fn test_decode_short_calldata() {
         // Test that short calldata returns proper error
-        let result = EthereumDecoder::decode_call_internal("0x1234", None);  // Only 2 bytes, need 4
+        let result = EthereumDecoder::decode_call_internal("0x1234", None); // Only 2 bytes, need 4
         assert!(result.is_err());
         assert!(result.unwrap_err().to_string().contains("Calldata too short"));
     }
