+++ gemstone/src/ethereum/mod.rs
@@ -3,4 +3,4 @@ pub mod decoder;
 pub mod jsonrpc;
 pub mod model;
 
-pub use decoder::{DecodedCall, DecodedCallParam, EthereumDecoder};
+pub use decoder::{GemDecodedCall, GemDecodedCallParam, EthereumDecoder};
